requirements.txt
Django==3.1.4
pytest==6.1.2
pytest-django==4.1.0
buttercms-python==1.2
